requirements.txt
dill==0.3.8
inquirerpy==0.3.4
multiprocess==0.70.14
nestd==0.3.1
orjson==3.9.15
packaging==24.1
pfzy==0.3.4
prompt_toolkit==3.0.47
python-dotenv==1.0.1
robyn==0.56.2
rustimport==1.3.4
toml==0.10.2
uvloop==0.19.0
watchdog==2.2.1
wcwidth==0.2.13
